BZ1176265 ISPN-3959 JdbcBinaryStore's expiration locks buckets indefinitely
ISPN-4520 JdbcBinaryStoreTest.testLoadAndStoreWithLifespanAndIdle random failures

* Increased the timeout to avoid false positives

ISPN-3840 Test suite should continue when a test hasn't closed its cache managers

Add fail-all-tests and fail-all-tests-setup profile to fail all the tests
and @Before methods and check that their cache managers are closed properly.

ISPN-4580 Refactor BaseStoreTest

...and its subclasses

* Remove duplicated code
* replace assert keyword by AssertJUnit methods
* use a mock cache in the stores to test

Delete jpastoretest

ISPN-4581 Memory leak when H2 is used

* clear the cache before killing the cache manager

ISPN-4452 DummyInMemoryStore should respect fetch arguments on process method.

ISPN-4520 JdbcBinaryStoreTest.testLoadAndStoreWithLifespanAndIdle random failures

* Use a custom TimeService to trigger the expiration

Statically imported method was shadowed by member method on newer compilers

ISPN-4651 LevelDB crashes JVM when stop() is called concurrently with write()

Added call for PurgeListener to LevelDBStore

+++ persistence/leveldb/src/main/java/org/infinispan/persistence/leveldb/LevelDBStore.java
@@ -1,5 +1,18 @@
 package org.infinispan.persistence.leveldb;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.Callable;
+import java.util.concurrent.CompletionService;
+import java.util.concurrent.Executor;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.Semaphore;
+
 import org.infinispan.commons.CacheConfigurationException;
 import org.infinispan.commons.configuration.ConfiguredBy;
 import org.infinispan.commons.util.Util;
@@ -16,23 +29,10 @@ import org.infinispan.persistence.spi.PersistenceException;
 import org.infinispan.util.logging.LogFactory;
 import org.iq80.leveldb.CompressionType;
 import org.iq80.leveldb.DB;
-import org.iq80.leveldb.DBException;
 import org.iq80.leveldb.DBFactory;
 import org.iq80.leveldb.DBIterator;
 import org.iq80.leveldb.Options;
 import org.iq80.leveldb.ReadOptions;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.Callable;
-import java.util.concurrent.CompletionService;
-import java.util.concurrent.Executor;
-import java.util.concurrent.LinkedBlockingQueue;
 
 /**
  * @public
@@ -51,13 +51,15 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
    private DB db;
    private DB expiredDb;
    private InitializationContext ctx;
-
+   private Semaphore semaphore;
+   private volatile boolean stopped = true;
 
    @Override
    public void init(InitializationContext ctx) {
       this.configuration = ctx.getConfiguration();
       this.dbFactory = newDbFactory();
       this.ctx = ctx;
+      this.semaphore = new Semaphore(Integer.MAX_VALUE, true);
 
       if (this.dbFactory == null) {
          throw log.cannotLoadlevelDBFactories(Arrays.toString(DB_FACTORY_CLASS_NAMES));
@@ -99,6 +101,7 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
       try {
          db = openDatabase(getQualifiedLocation(), dataDbOptions());
          expiredDb = openDatabase(getQualifiedExpiredLocation(), expiredDbOptions());
+         stopped = false;
       } catch (IOException e) {
          throw new CacheConfigurationException("Unable to open database", e);
       }
@@ -159,64 +162,95 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
    }
 
    protected void reinitAllDatabases() throws IOException {
-      // assuming this is only called from within clearLockSafe()
-      // clear() acquires a global lock before calling clearLockSafe()
       try {
-         db.close();
-      } catch (IOException e) {
-         log.warnUnableToCloseDb(e);
+         semaphore.acquire(Integer.MAX_VALUE);
+      } catch (InterruptedException e) {
+         throw new PersistenceException("Cannot acquire semaphore", e);
       }
       try {
-         expiredDb.close();
-      } catch (IOException e) {
-         log.warnUnableToCloseExpiredDb(e);
+         if (stopped) {
+            throw new PersistenceException("LevelDB is stopped");
+         }
+         try {
+            db.close();
+         } catch (IOException e) {
+            log.warnUnableToCloseDb(e);
+         }
+         try {
+            expiredDb.close();
+         } catch (IOException e) {
+            log.warnUnableToCloseExpiredDb(e);
+         }
+         db = reinitDatabase(getQualifiedLocation(), dataDbOptions());
+         expiredDb = reinitDatabase(getQualifiedExpiredLocation(), expiredDbOptions());
+      } finally {
+         semaphore.release(Integer.MAX_VALUE);
       }
-      db = reinitDatabase(getQualifiedLocation(), dataDbOptions());
-      expiredDb = reinitDatabase(getQualifiedExpiredLocation(), expiredDbOptions());
    }
 
    @Override
    public void stop()  {
       try {
-         db.close();
-      } catch (IOException e) {
-         log.warnUnableToCloseDb(e);
+         semaphore.acquire(Integer.MAX_VALUE);
+      } catch (InterruptedException e) {
+         throw new PersistenceException("Cannot acquire semaphore", e);
       }
-
       try {
-         expiredDb.close();
-      } catch (IOException e) {
-         log.warnUnableToCloseExpiredDb(e);
+         try {
+            db.close();
+         } catch (IOException e) {
+            log.warnUnableToCloseDb(e);
+         }
+
+         try {
+            expiredDb.close();
+         } catch (IOException e) {
+            log.warnUnableToCloseExpiredDb(e);
+         }
+      } finally {
+         stopped = true;
+         semaphore.release(Integer.MAX_VALUE);
       }
    }
 
    @Override
    public void clear() {
       long count = 0;
-      DBIterator it = db.iterator(new ReadOptions().fillCache(false));
       boolean destroyDatabase = false;
+      try {
+         semaphore.acquire();
+      } catch (InterruptedException e) {
+         throw new PersistenceException("Cannot acquire semaphore", e);
+      }
+      try {
+         if (stopped) {
+            throw new PersistenceException("LevelDB is stopped");
+         }
+         DBIterator it = db.iterator(new ReadOptions().fillCache(false));
+         if (configuration.clearThreshold() <= 0) {
+            try {
+               for (it.seekToFirst(); it.hasNext(); ) {
+                  Map.Entry<byte[], byte[]> entry = it.next();
+                  db.delete(entry.getKey());
+                  count++;
 
-      if (configuration.clearThreshold() <= 0) {
-         try {
-            for (it.seekToFirst(); it.hasNext();) {
-               Map.Entry<byte[], byte[]> entry = it.next();
-               db.delete(entry.getKey());
-               count++;
-
-               if (count > configuration.clearThreshold()) {
-                  destroyDatabase = true;
-                  break;
+                  if (count > configuration.clearThreshold()) {
+                     destroyDatabase = true;
+                     break;
+                  }
+               }
+            } finally {
+               try {
+                  it.close();
+               } catch (IOException e) {
+                  log.warnUnableToCloseDbIterator(e);
                }
             }
-         } finally {
-            try {
-               it.close();
-            } catch (IOException e) {
-               log.warnUnableToCloseDbIterator(e);
-            }
+         } else {
+            destroyDatabase = true;
          }
-      } else {
-         destroyDatabase = true;
+      } finally {
+         semaphore.release();
       }
 
       if (destroyDatabase) {
@@ -251,33 +285,45 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
       final TaskContext taskContext = new TaskContextImpl();
 
       List<Map.Entry<byte[], byte[]>> entries = new ArrayList<Map.Entry<byte[], byte[]>>(batchSize);
-      DBIterator it = db.iterator(new ReadOptions().fillCache(false));
       try {
-         for (it.seekToFirst(); it.hasNext();) {
-            Map.Entry<byte[], byte[]> entry = it.next();
-            entries.add(entry);
-            if (entries.size() == batchSize) {
-               final List<Map.Entry<byte[], byte[]>> batch = entries;
-               entries = new ArrayList<Map.Entry<byte[], byte[]>>(batchSize);
-               submitProcessTask(cacheLoaderTask, keyFilter,eacs, taskContext, batch, loadValues, loadMetadata);
+         semaphore.acquire();
+      } catch (InterruptedException e) {
+         throw new PersistenceException("Cannot acquire semaphore: CacheStore is likely stopped.", e);
+      }
+      try {
+         if (stopped) {
+            throw new PersistenceException("LevelDB is stopped");
+         }
+         DBIterator it = db.iterator(new ReadOptions().fillCache(false));
+         try {
+            for (it.seekToFirst(); it.hasNext(); ) {
+               Map.Entry<byte[], byte[]> entry = it.next();
+               entries.add(entry);
+               if (entries.size() == batchSize) {
+                  final List<Map.Entry<byte[], byte[]>> batch = entries;
+                  entries = new ArrayList<Map.Entry<byte[], byte[]>>(batchSize);
+                  submitProcessTask(cacheLoaderTask, keyFilter, eacs, taskContext, batch, loadValues, loadMetadata);
+               }
+            }
+            if (!entries.isEmpty()) {
+               submitProcessTask(cacheLoaderTask, keyFilter, eacs, taskContext, entries, loadValues, loadMetadata);
+            }
+
+            eacs.waitUntilAllCompleted();
+            if (eacs.isExceptionThrown()) {
+               throw new PersistenceException("Execution exception!", eacs.getFirstException());
+            }
+         } catch (Exception e) {
+            throw new PersistenceException(e);
+         } finally {
+            try {
+               it.close();
+            } catch (IOException e) {
+               log.warnUnableToCloseDbIterator(e);
             }
          }
-         if (!entries.isEmpty()) {
-            submitProcessTask(cacheLoaderTask, keyFilter,eacs, taskContext, entries, loadValues, loadMetadata);
-         }
-
-         eacs.waitUntilAllCompleted();
-         if (eacs.isExceptionThrown()) {
-            throw new PersistenceException("Execution exception!", eacs.getFirstException());
-         }
-      } catch (Exception e) {
-         throw new PersistenceException(e);
       } finally {
-         try {
-            it.close();
-         } catch (IOException e) {
-            log.warnUnableToCloseDbIterator(e);
-         }
+         semaphore.release();
       }
    }
 
@@ -318,10 +364,18 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
    public boolean delete(Object key)  {
       try {
          byte[] keyBytes = marshall(key);
-         if (db.get(keyBytes) == null) {
-            return false;
+         semaphore.acquire();
+         try {
+            if (stopped) {
+               throw new PersistenceException("LevelDB is stopped");
+            }
+            if (db.get(keyBytes) == null) {
+               return false;
+            }
+            db.delete(keyBytes);
+         } finally {
+            semaphore.release();
          }
-         db.delete(keyBytes);
          return true;
       } catch (Exception e) {
          throw new PersistenceException(e);
@@ -331,20 +385,40 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
    @Override
    public void write(MarshalledEntry me)  {
       try {
-         db.put(marshall(me.getKey()), marshall(me));
+         byte[] marshelledKey = marshall(me.getKey());
+         byte[] marshalledEntry = marshall(me);
+         semaphore.acquire();
+         try {
+            if (stopped) {
+               throw new PersistenceException("LevelDB is stopped");
+            }
+            db.put(marshelledKey, marshalledEntry);
+         } finally {
+            semaphore.release();
+         }
          InternalMetadata meta = me.getMetadata();
          if (meta != null && meta.expiryTime() > -1) {
             addNewExpiry(me);
          }
       } catch (Exception e) {
-         throw new DBException(e);
+         throw new PersistenceException(e);
       }
    }
 
    @Override
    public MarshalledEntry load(Object key)  {
       try {
-         MarshalledEntry me = (MarshalledEntry) unmarshall(db.get(marshall(key)));
+         byte[] marshalledEntry;
+         semaphore.acquire();
+         try {
+            if (stopped) {
+               throw new PersistenceException("LevelDB is stopped");
+            }
+            marshalledEntry = db.get(marshall(key));
+         } finally {
+            semaphore.release();
+         }
+         MarshalledEntry me = (MarshalledEntry) unmarshall(marshalledEntry);
          if (me == null) return null;
 
          InternalMetadata meta = me.getMetadata();
@@ -361,6 +435,14 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
    @Override
    public void purge(Executor executor, PurgeListener purgeListener) {
       try {
+         semaphore.acquire();
+      } catch (InterruptedException e) {
+         throw new PersistenceException("Cannot acquire semaphore: CacheStore is likely stopped.", e);
+      }
+      try {
+         if (stopped) {
+            throw new PersistenceException("LevelDB is stopped");
+         }
          // Drain queue and update expiry tree
          List<ExpiryEntry> entries = new ArrayList<ExpiryEntry>();
          expiryEntryQueue.drainTo(entries);
@@ -389,12 +471,13 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
          List<Long> times = new ArrayList<Long>();
          List<Object> keys = new ArrayList<Object>();
          DBIterator it = expiredDb.iterator(new ReadOptions().fillCache(false));
+         long now = ctx.getTimeService().wallClockTime();
          try {
             for (it.seekToFirst(); it.hasNext();) {
                Map.Entry<byte[], byte[]> entry = it.next();
 
                Long time = (Long) unmarshall(entry.getKey());
-               if (time > System.currentTimeMillis())
+               if (time > now)
                   break;
                times.add(time);
                Object key = unmarshall(entry.getValue());
@@ -411,7 +494,6 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
             if (!keys.isEmpty())
                log.debugf("purge (up to) %d entries", keys.size());
             int count = 0;
-            long currentTimeMillis = System.currentTimeMillis();
             for (Object key : keys) {
                byte[] keyBytes = marshall(key);
 
@@ -419,11 +501,14 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
                if (b == null)
                   continue;
                MarshalledEntry me = (MarshalledEntry) ctx.getMarshaller().objectFromByteBuffer(b);
-               if (me.getMetadata() != null && me.getMetadata().isExpired(ctx.getTimeService().wallClockTime())) {
+               // TODO race condition: the entry could be updated between the get and delete!
+               if (me.getMetadata() != null && me.getMetadata().isExpired(now)) {
                   // somewhat inefficient to FIND then REMOVE...
                   db.delete(keyBytes);
+                  purgeListener.entryPurged(key);
                   count++;
                }
+
             }
             if (count != 0)
                log.debugf("purged %d entries", count);
@@ -440,6 +525,8 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
          throw e;
       } catch (Exception e) {
          throw new PersistenceException(e);
+      } finally {
+         semaphore.release();
       }
    }
 
@@ -462,7 +549,7 @@ public class LevelDBStore implements AdvancedLoadWriteStore {
          // being a moving target
          // which could lead to unexpected results, hence, InternalCacheEntry
          // calls are required
-         expiry = maxIdle + System.currentTimeMillis();
+         expiry = maxIdle + ctx.getTimeService().wallClockTime();
       }
       Long at = expiry;
       Object key = entry.getKey();
